Fixed id generation for strings

+++ src/persistence/IdentifiableSqlServerPersistence.ts
@@ -152,6 +152,9 @@ export class IdentifiableSqlServerPersistence<T extends IIdentifiable<K>, K> ext
      * @returns the requested data item or <code>null</code> if nothing was found.
      */
     public async getOneById(correlationId: string, id: K): Promise<T> {
+        if (this.isEmpty(id))
+            return null;
+        
         let query = "SELECT * FROM " + this.quotedTableName() + " WHERE [id]=@1";
         let params = [ id ];
 
@@ -191,7 +194,7 @@ export class IdentifiableSqlServerPersistence<T extends IIdentifiable<K>, K> ext
 
         // Assign unique id
         let newItem: any = item;
-        if (newItem.id == null && this._autoGenerateId) {
+        if (this.isEmpty(newItem.id) && this._autoGenerateId) {
             newItem = Object.assign({}, newItem);
             newItem.id = item.id || IdGenerator.nextLong();
         }
@@ -213,7 +216,7 @@ export class IdentifiableSqlServerPersistence<T extends IIdentifiable<K>, K> ext
         }
 
         // Assign unique id
-        if (item.id == null && this._autoGenerateId) {
+        if (this.isEmpty(item.id) && this._autoGenerateId) {
             item = Object.assign({}, item);
             item.id = <any>IdGenerator.nextLong();
         }
@@ -283,7 +286,7 @@ export class IdentifiableSqlServerPersistence<T extends IIdentifiable<K>, K> ext
      * @returns the updated item.
      */
     public async update(correlationId: string, item: T): Promise<T> {
-        if (item == null || item.id == null) {
+        if (item == null || this.isEmpty(item.id)) {
             return null;
         }
 
@@ -323,7 +326,7 @@ export class IdentifiableSqlServerPersistence<T extends IIdentifiable<K>, K> ext
      * @returns the updated item.
      */
     public async updatePartially(correlationId: string, id: K, data: AnyValueMap): Promise<T> {
-        if (data == null || id == null) {
+        if (data == null || this.isEmpty(id)) {
             return null;
         }
 
@@ -362,6 +365,8 @@ export class IdentifiableSqlServerPersistence<T extends IIdentifiable<K>, K> ext
      * @returns the deleted item.
      */
     public async deleteById(correlationId: string, id: K): Promise<T> {
+        if (this.isEmpty(id))
+            return null;
         let values = [ id ];
 
         let query = "DELETE FROM " + this.quotedTableName() + " OUTPUT DELETED.* WHERE [id]=@1";
@@ -408,5 +413,21 @@ export class IdentifiableSqlServerPersistence<T extends IIdentifiable<K>, K> ext
         });
 
         this._logger.trace(correlationId, "Deleted %d items from %s", count, this._tableName);
+    }
+
+    /**
+     * Checks if value is empty
+     * @param value any value
+     * @returns true if value empty, other false
+     */
+    protected isEmpty(value: any) {
+        const type = typeof value;
+        if (value !== null && type === 'object' || type === 'function') {
+            const props = Object.keys(value);
+                if (props.length === 0) { 
+                    return true;
+                } 
+            } 
+        return !value;
     }
 }
